requirements.txt
django
psycopg2
south
dj_database_url
yolk
gunicorn
django-storages
boto
django-tastypie
requests
django-celery
collectfast
django-custom-user
django-oauth2-provider
django-haystack
selenium
mock